fix: Bucket Calculator improvements and form cleanup
1. Allow decimal ages (e.g., 57.5) for bucket start/end ages:
   - Changed WithdrawalBucket model fields from IntegerField to DecimalField
   - Added decimal_places=1 to allow tenths (e.g., 57.5)
   - Updated form widgets to include step="0.1"
   - Created migration 0006 to alter field types

2. Removed unused form fields:
   - Removed Tax Loss Harvesting checkbox from bucket form template
   - Removed Roth Conversions checkbox from bucket form template
   - These features are not yet implemented, so hiding them reduces confusion

3. Chart dark mode labels:
   - Verified all charts already have dark mode support configured
   - Charts detect dark mode and use appropriate text/grid colors
   - Monte Carlo, Bucket, and other charts all use textColor/gridColor variables

Changes:
- jretirewise/scenarios/models.py: Changed start_age/end_age to DecimalField
- jretirewise/scenarios/forms.py: Added step="0.1" to age input widgets
- jretirewise/scenarios/migrations/0006_*.py: Migration for field type changes
- jretirewise/templates/jretirewise/bucket_form.html: Removed Advanced Options section

+++ jretirewise/scenarios/models.py
@@ -117,8 +117,8 @@ class WithdrawalBucket(models.Model):
     order = models.IntegerField(default=0)
 
     # Time period (age range OR year range)
-    start_age = models.IntegerField(null=True, blank=True)
-    end_age = models.IntegerField(null=True, blank=True)
+    start_age = models.DecimalField(max_digits=5, decimal_places=1, null=True, blank=True)
+    end_age = models.DecimalField(max_digits=5, decimal_places=1, null=True, blank=True)
     start_year = models.IntegerField(null=True, blank=True)
     end_year = models.IntegerField(null=True, blank=True)
